style: 슬랙 로직 추상화한 클래스 이름 변경
- 인터페이스 SlackClient -> ExternalCilent
- 실제 구현체 SlackClientIml -> SlackClient
- 테스트용 구현체 SlackClientTest -> FakeClient
