+++ backend/src/main/java/com/pickpick/member/application/MemberInitializer.java
@@ -2,7 +2,7 @@ package com.pickpick.member.application;
 
 import com.pickpick.member.domain.Member;
 import com.pickpick.member.domain.MemberRepository;
-import com.pickpick.support.SlackClient;
+import com.pickpick.support.ExternalClient;
 import java.util.List;
 import java.util.stream.Collectors;
 import javax.annotation.PostConstruct;
@@ -13,10 +13,10 @@ import org.springframework.stereotype.Component;
 @Component
 public class MemberInitializer {
 
-    private final SlackClient slackClient;
+    private final ExternalClient slackClient;
     private final MemberRepository memberRepository;
 
-    public MemberInitializer(final SlackClient slackClient, final MemberRepository memberRepository) {
+    public MemberInitializer(final ExternalClient slackClient, final MemberRepository memberRepository) {
         this.slackClient = slackClient;
         this.memberRepository = memberRepository;
     }
