+++ backend/src/main/java/com/pickpick/message/application/ReminderSender.java
@@ -2,7 +2,7 @@ package com.pickpick.message.application;
 
 import com.pickpick.message.domain.Reminder;
 import com.pickpick.message.domain.ReminderRepository;
-import com.pickpick.support.SlackClient;
+import com.pickpick.support.ExternalClient;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
@@ -16,9 +16,9 @@ import org.springframework.transaction.annotation.Transactional;
 public class ReminderSender {
 
     private final ReminderRepository reminders;
-    private final SlackClient slackClient;
+    private final ExternalClient slackClient;
 
-    public ReminderSender(final ReminderRepository reminders, final SlackClient slackClient) {
+    public ReminderSender(final ReminderRepository reminders, final ExternalClient slackClient) {
         this.reminders = reminders;
         this.slackClient = slackClient;
     }
